+++ app/containers/MessagingContentArea/index.js
@@ -217,7 +217,7 @@ export class MessagingContentArea extends React.Component {
       );
     }
 
-    return <ContentArea from={from} buttons={buttons} details={details} content={content} />;
+    return <ContentArea interaction={this.props.selectedInteraction} from={from} buttons={buttons} details={details} content={content} />;
   }
 }