+++ Changelog.md
@@ -1,3 +1,13 @@
+<a name"0.3.1"></a>
+### 0.3.1 (2017-02-23)
+* Fix contacts panel size on window resize
+* Added 'active' class names for voice buttons
+
+<a name"0.3.0"></a>
+### 0.3.0 (2017-02-23)
+* Notes components
+* Updated local NovaX (native messaging) setup
+
 <a name"0.2.25"></a>
 ### 0.2.25 (2017-02-22)
 * fix FB name on messages